Use a DialogFragment to confirm the deletion of a book
Showing an AlertDialog directly from a fragment is discouraged. It
doesn't behave properly, for example, when rotating the screen. See
https://developer.android.com/guide/topics/ui/dialogs.html
