+++ app/src/main/java/org/gnucash/android/ui/settings/BookManagerFragment.java
@@ -38,8 +38,6 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CheckBox;
-import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.ListView;
@@ -56,6 +54,7 @@ import org.gnucash.android.db.adapter.SplitsDbAdapter;
 import org.gnucash.android.db.adapter.TransactionsDbAdapter;
 import org.gnucash.android.ui.account.AccountsActivity;
 import org.gnucash.android.ui.common.Refreshable;
+import org.gnucash.android.ui.settings.dialog.DeleteBookConfirmationDialog;
 import org.gnucash.android.util.BookUtils;
 import org.gnucash.android.util.PreferencesHelper;
 
@@ -219,34 +218,9 @@ public class BookManagerFragment extends ListFragment implements
         }
 
         private boolean handleMenuDeleteBook(final String bookUID) {
-            AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(getActivity());
-            dialogBuilder.setTitle(getString(R.string.title_confirm_delete_book))
-                    .setIcon(R.drawable.ic_close_black_24dp)
-                    .setView(R.layout.dialog_double_confirm)
-                    .setMessage(getString(R.string.msg_all_book_data_will_be_deleted));
-            dialogBuilder.setPositiveButton(getString(R.string.btn_delete_book), new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    BooksDbAdapter.getInstance().deleteBook(bookUID);
-                    refresh();
-                }
-            });
-            dialogBuilder.setNegativeButton(R.string.btn_cancel, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    dialog.dismiss();
-                }
-            });
-            final AlertDialog dialog = dialogBuilder.create();
-            dialog.show(); //must be called before you can access buttons
-            dialog.getButton(AlertDialog.BUTTON_POSITIVE).setEnabled(false);
-            CheckBox confirmCheckBox = dialog.findViewById(R.id.checkbox_confirm);
-            confirmCheckBox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                @Override
-                public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
-                    dialog.getButton(AlertDialog.BUTTON_POSITIVE).setEnabled(b);
-                }
-            });
+            DeleteBookConfirmationDialog dialog = DeleteBookConfirmationDialog.newInstance(bookUID);
+            dialog.show(getFragmentManager(), "delete_book");
+            dialog.setTargetFragment(BookManagerFragment.this, 0);
             return true;
         }
